refactor/all: split file into smaller ones
This splits lib.rs into three new files.
